同步 Stockfish 最新的版本
合入的差异部分未 Stockfish
2020/5/8 的提交 66ed8b6c479932f1ec2274b5f567b5a6aecae0a4 到
2020/12/8 的提交 d706ae62d73d90c0f80cdccd58384a347295d549 的差异部分。

暂时未使用的代码不合并。
nnue 主体未合并。
main.cpp() 中 CommandLine::init(argc, argv); 未合并。

+++ src/main.cpp
@@ -49,7 +49,7 @@ int main(int argc, char* argv[]) {
   Position::init();
   //Bitbases::init();
   //Endgames::init();
-  Threads.set((int)Options["Threads"]);
+  Threads.set(size_t(Options["Threads"]));
   Search::clear(); // After threads are up
 
   UCI::loop(argc, argv);
